L2J Public release No. 1 Codename - Informative Idiotus Changelog will be possibly posted later (changes are self-explanatory) Enjoy. Next release scheduled to 30/11/2010 16:34 CET.

+++ datapack_development/sql/skill_residential.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS `skill_residential`;
 CREATE TABLE `skill_residential` (
-  `entityId` int(11) NOT NULL,
-  `skillId` int(11) NOT NULL DEFAULT '0',
-  `skillLevel` int(11) NOT NULL DEFAULT '0',
+  `entityId` tinyint(3) unsigned NOT NULL,
+  `skillId` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `skillLevel` tinyint(2) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`entityId`,`skillId`)
 );
